Reorganize to have auth related code in auth/*
