+++ client/actions/groups/group_client.go
@@ -2,9 +2,9 @@ package groups
 
 import (
 	"errors"
+	"github.com/IBM/satcon-client-go/client/auth"
 	"net/http"
 
-	"github.com/IBM/satcon-client-go/client/actions"
 	"github.com/IBM/satcon-client-go/client/types"
 	"github.com/IBM/satcon-client-go/client/web"
 )
@@ -28,7 +28,7 @@ type Client struct {
 
 // NewClient returns a configured instance of GroupService which can then be used
 // to perform group queries against Satellite Config.
-func NewClient(endpointURL string, httpClient web.HTTPClient, authClient actions.AuthClient) (GroupService, error) {
+func NewClient(endpointURL string, httpClient web.HTTPClient, authClient auth.AuthClient) (GroupService, error) {
 	if endpointURL == "" {
 		return nil, errors.New("Must supply a valid endpoint URL")
 	}
