+++ client/actions/resources/resource_client.go
@@ -2,9 +2,9 @@ package resources
 
 import (
 	"errors"
+	"github.com/IBM/satcon-client-go/client/auth"
 	"net/http"
 
-	"github.com/IBM/satcon-client-go/client/actions"
 	"github.com/IBM/satcon-client-go/client/types"
 	"github.com/IBM/satcon-client-go/client/web"
 )
@@ -25,7 +25,7 @@ type Client struct {
 
 // NewClient returns a configured instance of ClusterService which can then be used
 // to perform cluster queries against Satellite Config.
-func NewClient(endpointURL string, httpClient web.HTTPClient, authClient actions.AuthClient) (ResourceService, error) {
+func NewClient(endpointURL string, httpClient web.HTTPClient, authClient auth.AuthClient) (ResourceService, error) {
 	if endpointURL == "" {
 		return nil, errors.New("Must supply a valid endpoint URL")
 	}
