+++ client/actions/versions/version_client_test.go
@@ -1,6 +1,7 @@
 package versions_test
 
 import (
+	"github.com/IBM/satcon-client-go/client/auth/iam"
 	"net/http"
 	"time"
 
@@ -8,7 +9,6 @@ import (
 	. "github.com/onsi/gomega"
 
 	. "github.com/IBM/satcon-client-go/client/actions/versions"
-	"github.com/IBM/satcon-client-go/client/auth"
 )
 
 var _ = Describe("ClusterClient", func() {
@@ -16,13 +16,13 @@ var _ = Describe("ClusterClient", func() {
 		var (
 			h         *http.Client
 			endpoint  string
-			iamClient *auth.IAMClient
+			iamClient *iam.IAMClient
 			err       error
 		)
 
 		BeforeEach(func() {
 			endpoint = "https://satcon.foo"
-			iamClient, err = auth.NewIAMClient("some_key")
+			iamClient, err = iam.NewIAMClient("some_key")
 			Expect(err).ToNot(HaveOccurred())
 
 		})
