+++ client/auth/client.go
@@ -1,105 +1,8 @@
 package auth
 
-import (
-	"fmt"
-	"github.com/IBM/satcon-client-go/client/actions"
-	"github.com/IBM/satcon-client-go/client/auth/local"
-	"github.com/IBM/satcon-client-go/client/types"
-	"github.com/IBM/satcon-client-go/client/web"
-	"github.com/dgrijalva/jwt-go"
-	"net/http"
-	"time"
+import "net/http"
 
-	"github.com/IBM/go-sdk-core/v4/core"
-)
-
-//IAMClient manages authorization for Satcon Client requests
-type IAMClient struct {
-	Client actions.AuthClient
-}
-
-//New returns a new core.IamAuthenticator struct and also returns the error
-func NewIAMClient(apiKey string) (*IAMClient, error) {
-
-	iamClient, err := core.NewIamAuthenticator(apiKey, "", "", "", false, nil)
-
-	if err == nil {
-		return &IAMClient{Client: iamClient}, nil
-	}
-
-	return nil, err
-
-}
-
-const TokenValidityDuration = 40 * time.Minute
-const MinimumTimeTokenStillValid = 5 * time.Minute
-const AuthorizationHeaderKey = "Authorization"
-
-type LocalRazeeClient struct {
-	HTTPClient web.HTTPClient
-	url        string
-	login      string
-	password   string
-
-	// cached jwt token from previous request
-	token types.Token
-	// timestamp of the time when the previous
-	// cached token was retrieved
-	tokenTimestamp time.Time
-	// timestamp that was set in the 'exp' claim of
-	// the previous retrieved jwt token.
-	expireTimestamp time.Time
-}
-
-func (l *LocalRazeeClient) Authenticate(request *http.Request) error {
-	invalidExpiredTimestamp := time.Until(l.expireTimestamp) < MinimumTimeTokenStillValid
-	invalidTokenTimestamp := time.Since(l.tokenTimestamp) >= TokenValidityDuration
-	if l.token == "" || invalidExpiredTimestamp || invalidTokenTimestamp {
-		token, err := local.SignIn(l.HTTPClient, l.url, l.login, l.password)
-		if err != nil {
-			return err
-		}
-		if token == nil {
-			return fmt.Errorf("Could not get a token by signing in %v to %v", l.login, l.url)
-		}
-		l.token = *token
-		parsedToken, err := jwt.Parse(string(*token), nil)
-		if parsedToken != nil {
-			claims, _ := parsedToken.Claims.(jwt.MapClaims)
-			if expiredTimestamp, ok := claims["exp"]; ok {
-				if d, ok := expiredTimestamp.(float64); ok {
-					l.expireTimestamp = time.Unix(int64(d), 0)
-				}
-			}
-		}
-
-		// Token timestamp is a backup mechanism if the token does not contain the 'exp' field
-		// or the value of that field could not be parsed
-		l.tokenTimestamp = time.Now()
-	}
-	request.Header.Add(AuthorizationHeaderKey, "Bearer "+string(l.token))
-	return nil
-}
-
-func NewLocalRazeeClient(url string, login string, password string) (*LocalRazeeClient, error) {
-	return NewLocalRazeeClientWithHttpClient(http.DefaultClient, url, login, password)
-}
-
-func NewLocalRazeeClientWithHttpClient(httpClient web.HTTPClient, url string, login string, password string) (*LocalRazeeClient, error) {
-	if url == "" {
-		return nil, fmt.Errorf("Field 'url' cannot be empty!")
-	}
-	if login == "" {
-		return nil, fmt.Errorf("Field 'login' cannot be empty!")
-	}
-	if password == "" {
-		return nil, fmt.Errorf("Field 'password' cannot be empty!")
-	}
-	client := &LocalRazeeClient{
-		HTTPClient: httpClient,
-		url:        url,
-		login:      login,
-		password:   password,
-	}
-	return client, nil
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . AuthClient
+type AuthClient interface {
+	Authenticate(request *http.Request) error
 }
