+++ client/client_test.go
@@ -1,6 +1,7 @@
 package client_test
 
 import (
+	"github.com/IBM/satcon-client-go/client/auth/iam"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
@@ -12,20 +13,19 @@ import (
 	"github.com/IBM/satcon-client-go/client/actions/resources/resourcesfakes"
 	"github.com/IBM/satcon-client-go/client/actions/subscriptions/subscriptionsfakes"
 	"github.com/IBM/satcon-client-go/client/actions/versions/versionsfakes"
-	"github.com/IBM/satcon-client-go/client/auth"
 )
 
 var _ = Describe("Client", func() {
 	Describe("New", func() {
 		var (
 			endpointURL string
-			iamClient   *auth.IAMClient
+			iamClient   *iam.IAMClient
 			err         error
 		)
 
 		BeforeEach(func() {
 			endpointURL = "https://foo.bar"
-			iamClient, err = auth.NewIAMClient("some_key")
+			iamClient, err = iam.NewIAMClient("some_key")
 			Expect(err).NotTo(HaveOccurred())
 		})
