+++ test/integration/groups_test.go
@@ -2,13 +2,13 @@ package integration_test
 
 import (
 	"fmt"
+	"github.com/IBM/satcon-client-go/client/auth/iam"
 	"strings"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/IBM/satcon-client-go/client"
-	"github.com/IBM/satcon-client-go/client/auth"
 	. "github.com/IBM/satcon-client-go/test/integration"
 )
 
@@ -16,12 +16,12 @@ var _ = Describe("Groups", func() {
 
 	var (
 		c         client.SatCon
-		iamClient *auth.IAMClient
+		iamClient *iam.IAMClient
 	)
 
 	BeforeEach(func() {
 		var err error
-		iamClient, err = auth.NewIAMClient(testConfig.APIKey)
+		iamClient, err = iam.NewIAMClient(testConfig.APIKey)
 		Expect(err).ToNot(HaveOccurred())
 		c, _ = client.New(testConfig.SatConEndpoint, nil, iamClient.Client)
 		Expect(c.Groups).NotTo(BeNil())
