+++ test/integration/users_test.go
@@ -2,23 +2,23 @@ package integration_test
 
 import (
 	"fmt"
+	"github.com/IBM/satcon-client-go/client/auth/iam"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/IBM/satcon-client-go/client"
-	"github.com/IBM/satcon-client-go/client/auth"
 )
 
 var _ = Describe("Users", func() {
 	var (
 		c         client.SatCon
-		iamClient *auth.IAMClient
+		iamClient *iam.IAMClient
 	)
 
 	BeforeEach(func() {
 		var err error
-		iamClient, err = auth.NewIAMClient(testConfig.APIKey)
+		iamClient, err = iam.NewIAMClient(testConfig.APIKey)
 		Expect(err).ToNot(HaveOccurred())
 		c, _ = client.New(testConfig.SatConEndpoint, nil, iamClient.Client)
 		Expect(c.Clusters).NotTo(BeNil())
